Auto-elevate installprofile as a service
The way ConfigurationProfile works is by spawning some XPC "helper"
processes, who will be in charge of applying the actual payloads. Those
helpers can be found in the following folder:

/System/Library/PrivateFrameworks/ConfigurationProfiles.framework/XPCServices

Unfortunately, ConfigurationProfile spawns those helpers in the current
session - which means that even if we run installprofile with sudo, the
spawned helpers will run under that user, and be limited by its
permissions. This caused TCCProfileService to fail, as a normal user
doesn't have the (unix) privileges to modify files in the
/Library/Application Support/com.apple.TCC folder.

We fix this by elevating installprofile to root through a launchd
service. It creates a launchd service that respawns itself with a
special flag (indicating we're running as service), in the correct
folder, and targeting the right file. It then proceeds to load it in
launchd, uses launchd debug to redirect the service's stdout/stderr to
the current terminal, and finally starts it.

Afterwards, it's going to monitor the current processes to figure out
when the installprofile service stops, kill the launchd debug session,
and exit.

Note that the way it works is cursed and there is _surely_ a better way
to do this, but I couldn't find one, and this works well enough.